Move CollabEditor to components
